# static/open-api-specification.yaml
openapi: 3.0.2
info:
  title: REST API for the lifap5-backend-2019-2020
  description: The yaml description of the API, lively available at http://lifap5.univ-lyon1.fr/
  version: "0.1.1"
  contact:
    email: romuald.thion@univ-lyon1.fr
    name: Romuald THION
  license:
    name: CC BY-NC-SA 3.0 FR
    url: 'https://creativecommons.org/licenses/by-nc-sa/3.0/fr/'
servers:
  - url: 'http://localhost:3000/'
    description: A local development server
  - url: 'https://lifap5.univ-lyon1.fr/'
    description: The public production server
...
paths:
...
  /quizzes/:
    get:
      summary: All quizzes
      description: The list of all quizzes (TODO with pagination)
      operationId: getAllQuizzesHandler
      tags:
        - public
        - quizzes
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: 
                type: array
                items:
                  type: object
                  $ref: '#/components/schemas/ObjectQuiz'
    post:
      summary: Create a new quiz
      description: Create a quiz, in the database. quiz_id is auto generated, title and description are mandatory
      operationId: postQuizHandler
      tags:
        - authenticated
        - quizzes
      security:
        - XApiKeyAuth: []
      requestBody:
        description: The quizz to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ObjectQuiz'
            example:
              title: "QCM de test"
              description: "Un QCM supplémentaire"
              open: false
      responses:
        '201':
          description: Content created, returns the quiz's id 
          content:
            application/json:
              schema: 
                type: object
                properties:
                  quiz_id:
                    type: integer
                    example: 42
  /quizzes/{quiz_id}:
    put:
      summary: Update an existing quiz
      description: attributes created_at, quiz_id and owner_id cannot be modified, owner_id must be current user
      operationId: putQuizHandler
      tags:
        - authenticated
        - quizzes
      security:
        - XApiKeyAuth: []
      requestBody:
        description: The the fields to update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ObjectQuiz'
            example:
              title: "Titre a mettre à jour"
              description: "Un QCM supplémentaire - descritiption modifiée"
              open: true
      responses:
        '200':
          description: Content updated, returns the whole quizz
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectQuiz'
      parameters:
        - $ref: '#/components/parameters/quizIdParam'
    delete:
      summary: Delete an existing quiz
      description: owner_id must be current user
      operationId: delQuizHandler
      tags:
        - authenticated
        - quizzes
      security:
        - XApiKeyAuth: []
      responses:
        '200':
          description: Content deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectQuiz'
      parameters:
        - $ref: '#/components/parameters/quizIdParam'
components:
  parameters:
    quizIdParam:
      name: quiz_id
      in: path
      description: ID of a quiz
      required: true
      schema:
        type: integer
      style: simple

# https://swagger.io/docs/specification/authentication/api-keys/
  securitySchemes:
    XApiKeyAuth:        # arbitrary name for the security scheme
      type: apiKey
      in: header       # can be "header", "query" or "cookie"
      name: X-API-KEY  # name of the header, query parameter or cookie
...
  schemas:
...
    ObjectQuiz:
      type: object
      required:
        - quiz_id
        - title
        - description
        - owner_id
      properties:
        quiz_id:
          description: value is auto generated by the database
          type: integer
          example: 1
        created_at:
          type: string
          format: date-time
          example: "2020-03-11T09:56:12.372Z"
        title:
          type: string
          example: "QCM LIFAP5 #1"
        description:
          type: string
          example: Des questions de JS et lambda calcul
        owner_id:
          type: string
          example: test.user
        questions_number:
          type: integer
          example: 2
        questions_ids:
          type: array
          items:
            type: integer
          example: [0,1]
        open:
          type: boolean
          example: false
        total_weight:
          type: integer
          example: 4
...